+++ scripts/generate/templates/Arrays.opts.js
